Test: Member CQS 분리에 따른 테스트 코드 재구성
