+++ src/test/java/team/unibusk/backend/domain/member/application/command/MemberCommandServiceTest.java
@@ -1,10 +1,9 @@
-package team.unibusk.backend.domain.member.application;
+package team.unibusk.backend.domain.member.application.command;
 
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import team.unibusk.backend.domain.member.application.dto.request.MemberNameUpdateServiceRequest;
-import team.unibusk.backend.domain.member.application.dto.response.MemberInfoResponse;
 import team.unibusk.backend.domain.member.application.dto.response.MemberNameUpdateResponse;
 import team.unibusk.backend.domain.member.domain.Member;
 import team.unibusk.backend.domain.member.domain.MemberRepository;
@@ -13,10 +12,10 @@ import team.unibusk.backend.global.support.UnitTestSupport;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
 
-class MemberServiceTest extends UnitTestSupport {
+class MemberCommandServiceTest extends UnitTestSupport {
 
     @InjectMocks
-    private MemberService memberService;
+    private MemberCommandService memberCommandService;
 
     @Mock
     private MemberRepository memberRepository;
@@ -30,29 +29,13 @@ class MemberServiceTest extends UnitTestSupport {
                 .build();
         given(memberRepository.findByMemberId(1L)).willReturn(member);
 
-        MemberNameUpdateResponse response = memberService.updateMemberName(
+        MemberNameUpdateResponse response = memberCommandService.updateMemberName(
                 new MemberNameUpdateServiceRequest(1L, "김철수")
         );
 
         assertThat(response.memberId()).isEqualTo(1L);
         assertThat(response.email()).isEqualTo("test@email.com");
         assertThat(response.name()).isEqualTo("김철수");
-    }
-
-    @Test
-    void 내_정보를_조회하면_회원_정보가_반환된다() {
-        Member member = Member.builder()
-                .id(1L)
-                .email("test@email.com")
-                .name("홍길동")
-                .build();
-        given(memberRepository.findByMemberId(1L)).willReturn(member);
-
-        MemberInfoResponse response = memberService.getMyInfo(1L);
-
-        assertThat(response.memberId()).isEqualTo(1L);
-        assertThat(response.email()).isEqualTo("test@email.com");
-        assertThat(response.name()).isEqualTo("홍길동");
     }
 
 }
